Use drawable->get/setType, log error detail
