+++ src/mbgl/renderer/layers/render_fill_layer.hpp
@@ -22,6 +22,15 @@ using FillLayerTweakerPtr = std::shared_ptr<FillLayerTweaker>;
 
 class RenderFillLayer final : public RenderLayer {
 public:
+    enum class FillVariant : uint8_t {
+        Fill,
+        FillPattern,
+        FillOutline,
+        FillOutlinePattern,
+        Undefined = 255
+    };
+    
+public:
     explicit RenderFillLayer(Immutable<style::FillLayer::Impl>);
     ~RenderFillLayer() override;
